+++ modules/home/screens/HomeScreen.tsx
@@ -87,7 +87,7 @@ export default function HomeScreen() {
         <View className="flex-row flex-wrap justify-between">
           <View className="w-[49%] mb-4">
             <ModuleButton
-              icon={emojis.approved}
+              //icon={emojis.approved}
               //icon={   emojis.package}
               label="Aprobación Pedidos"
               onPress={() =>
@@ -99,13 +99,13 @@ export default function HomeScreen() {
           <View className="w-[49%] mb-4">
             <ModuleButton
               //icon={emojis.list}
-              icon={emojis.approved}
+              //icon={emojis.approved}
               label="Consultar Pedidos"
               onPress={() => router.push("/(main)/(tabs)/(orders)/orderSearch")}
               bgColor="bg-secondary dark:bg-dark-secondary"
             />
           </View>
-          <View className="w-[49%] mb-4">
+          {/* <View className="w-[49%] mb-4">
             <ModuleButton
               //icon={emojis.bags}
               icon={emojis.approved}
@@ -113,7 +113,7 @@ export default function HomeScreen() {
               bgColor="bg-gray-300 dark:bg-gray-700"
               isComingSoon
             />
-          </View>
+          </View> */}
         </View>
       </View>
     </ScrollView>