+++ modules/orders/screens/OrderSearchScreen.tsx
@@ -98,8 +98,8 @@ export default function OrderSearchScreen() {
           handleRefresh={()=>handleRefresh(filters)}
           cooldown={cooldown}
           showtitle={true}
-          title={`${totalOrders} pedidos`}
-          subtitle={`con total ${totalVenezuela(totalUSD)}$`}
+          title={`${totalOrders} ${totalOrders > 1 ?  'pedidos': 'pedido'}`}
+          subtitle={`Total ${totalVenezuela(totalUSD)}$`}
           ListEmptyComponent={
             <View className="p-10 items-center">
               <Text className="text-foreground dark:text-dark-foreground">