+++ CHANGELOG.md
@@ -5,7 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.2] - 2025-10-08
+## [Unreleased]
+
+## [0.0.3]
+
+### Chaged
+
+- UI and styles in TitleText from CustomFlatList
+- Text for list in OrdersModule
+- buttons in HomeScreen
+
+### Fixed
+
+- The app crashes when the `pedido.co_art` is null, a validation check has been implemented.
+- Added validation for discount for product list in Orders Module `hasDiscount = discount !== "0" && discount.length > 0`
+
+## [0.0.2](https://github.com/JavierTaborda/AppFrigilux/releases/tag/v0.0.2) - 2025-10-08
 
 ### Added